Move ToolsTab under tools/
