+++ src/main/java/com/atlauncher/gui/LauncherFrame.java
@@ -52,7 +52,7 @@ import com.atlauncher.gui.tabs.PacksBrowserTab;
 import com.atlauncher.gui.tabs.ServersTab;
 import com.atlauncher.gui.tabs.SettingsTab;
 import com.atlauncher.gui.tabs.Tab;
-import com.atlauncher.gui.tabs.ToolsTab;
+import com.atlauncher.gui.tabs.tools.ToolsTab;
 import com.atlauncher.gui.tabs.VanillaPacksTab;
 import com.atlauncher.managers.AccountManager;
 import com.atlauncher.managers.PackManager;
